requirements.txt
streamlit
python-dotenv
langchain_openai
pinecone
openai